+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The `com.walmartlabs.lacinia.pedestal.interceptors' namespaces has been
 removed outright, and with it, support for interceptor dependency maps.
 A scattering of other functions related to dependency maps have also been removed.
 
+Dependencies have been updated to latest, including bundling of GraphiQL 0.11.11.
+
 ## 0.7.0 -- 2 Mar 2018
 
 The `com.walmartlabs.lacina.pedestal.interceptors` namespace has been deprecated; it will be removed