+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The `com.walmartlabs.lacinia.pedestal.interceptors' namespaces has been
 removed outright, and with it, support for interceptor dependency maps.
 A scattering of other functions related to dependency maps have also been removed.
 
+Added a number of options to control the creation of core.async channels used
+to manage the flow of data for subscriptions.
+
 ## 0.7.0 -- 2 Mar 2018
 
 The `com.walmartlabs.lacina.pedestal.interceptors` namespace has been deprecated; it will be removed