+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 0.8.0 -- UNRELEASED
 
-The `com.walmartlabs.lacinia.pedestal.interceptors' namespaces has been
+The `com.walmartlabs.lacinia.pedestal.interceptors` namespaces has been
 removed outright, and with it, support for interceptor dependency maps.
 A scattering of other functions related to dependency maps have also been removed.
+
+For subscription requests, the connection parameters are now captured at connection time,
+and available as the `:connection-params` key of the application context.
 
 ## 0.7.0 -- 2 Mar 2018