+++ CHANGELOG.md
@@ -10,7 +10,7 @@ to manage the flow of data for subscriptions.
 Dependencies have been updated to latest, including bundling of GraphiQL 0.11.11.
 
 For subscription requests, the connection parameters are now captured at connection time,
-and available as the `:connection-params` key of the application context.
+and available as the `:com.walmartlabs.lacinia/connection-params` key of the application context.
 
 ## 0.7.0 -- 2 Mar 2018