+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.8.0 -- UNRELEASED
+## 0.8.0 -- 6 Jun 2018
 
 The `com.walmartlabs.lacinia.pedestal.interceptors` namespaces has been
 removed outright, and with it, support for interceptor dependency maps.
@@ -11,6 +11,8 @@ Dependencies have been updated to latest.
 
 For subscription requests, the connection parameters are now captured at connection time,
 and available as the `:com.walmartlabs.lacinia/connection-params` key of the application context.
+
+[Closed Issues](https://github.com/walmartlabs/lacinia-pedestal/milestone/7?closed=1)
 
 ## 0.7.0 -- 2 Mar 2018